CSSUnitValueTyped can hold only Float values

+++ web/core/src/jsTest/kotlin/CSSUnitApiTests.kt
@@ -73,39 +73,39 @@ class CSSUnitApiTests {
 
     @Test
     fun postfixInvocation() {
-        4.number.assertStructure(CSSnumberValue(4), "number postfix")
-        4.percent.assertStructure(CSSpercentValue(4), "percent posfix")
+        4.number.assertStructure(CSSnumberValue(4.toFloat()), "number postfix")
+        4.percent.assertStructure(CSSpercentValue(4.toFloat()), "percent posfix")
 
-        4.em.assertStructure(CSSemValue(4), "em postfix")
-        4.ch.assertStructure(CSSchValue(4), "ch postfix")
+        4.em.assertStructure(CSSemValue(4.toFloat()), "em postfix")
+        4.ch.assertStructure(CSSchValue(4.toFloat()), "ch postfix")
 
-        4.cssRem.assertStructure(CSSremValue(4))
+        4.cssRem.assertStructure(CSSremValue(4.toFloat()))
 
-        4.vw.assertStructure(CSSvwValue(4),"vw postfix")
-        4.vh.assertStructure(CSSvhValue(4), "vh postfix")
+        4.vw.assertStructure(CSSvwValue(4.toFloat()),"vw postfix")
+        4.vh.assertStructure(CSSvhValue(4.toFloat()), "vh postfix")
 
-        4.vmin.assertStructure(CSSvminValue(4), "vmin postfix")
-        4.vmax.assertStructure(CSSvmaxValue(4), "vmax postfix")
-        4.cm.assertStructure(CSScmValue(4), "cm postfix")
-        4.mm.assertStructure(CSSmmValue(4), "mm postfix")
+        4.vmin.assertStructure(CSSvminValue(4.toFloat()), "vmin postfix")
+        4.vmax.assertStructure(CSSvmaxValue(4.toFloat()), "vmax postfix")
+        4.cm.assertStructure(CSScmValue(4.toFloat()), "cm postfix")
+        4.mm.assertStructure(CSSmmValue(4.toFloat()), "mm postfix")
 
-        4.pt.assertStructure(CSSptValue(4), "pt postfix")
-        4.pc.assertStructure(CSSpcValue(4), "pc postfix")
-        4.px.assertStructure(CSSpxValue(4), "px postfix")
+        4.pt.assertStructure(CSSptValue(4.toFloat()), "pt postfix")
+        4.pc.assertStructure(CSSpcValue(4.toFloat()), "pc postfix")
+        4.px.assertStructure(CSSpxValue(4.toFloat()), "px postfix")
 
-        4.deg.assertStructure(CSSdegValue(4), "deg postfix")
-        4.grad.assertStructure(CSSgradValue(4), "grad postfix")
-        4.rad.assertStructure(CSSradValue(4), "rad postfix")
-        4.turn.assertStructure(CSSturnValue(4), "turn postfix")
+        4.deg.assertStructure(CSSdegValue(4.toFloat()), "deg postfix")
+        4.grad.assertStructure(CSSgradValue(4.toFloat()), "grad postfix")
+        4.rad.assertStructure(CSSradValue(4.toFloat()), "rad postfix")
+        4.turn.assertStructure(CSSturnValue(4.toFloat()), "turn postfix")
 
-        4.s.assertStructure(CSSsValue(4), "s postfix")
-        4.ms.assertStructure(CSSmsValue(4), "ms postfix")
+        4.s.assertStructure(CSSsValue(4.toFloat()), "s postfix")
+        4.ms.assertStructure(CSSmsValue(4.toFloat()), "ms postfix")
 
-        4.dpi.assertStructure(CSSdpiValue(4), "dpi postfix")
-        4.dpcm.assertStructure(CSSdpcmValue(4), "dpcm postfix")
-        4.dppx.assertStructure(CSSdppxValue(4), "dppx postfix")
+        4.dpi.assertStructure(CSSdpiValue(4.toFloat()), "dpi postfix")
+        4.dpcm.assertStructure(CSSdpcmValue(4.toFloat()), "dpcm postfix")
+        4.dppx.assertStructure(CSSdppxValue(4.toFloat()), "dppx postfix")
 
-        4.fr.assertStructure(CSSfrValue(4), "fr postfix")
+        4.fr.assertStructure(CSSfrValue(4.toFloat()), "fr postfix")
     }
 
 }
